chore: const reference in util

+++ casbin/util/escape_assertion.cpp
@@ -41,8 +41,8 @@ std::string EscapeAssertion(std::string s) {
     for (std::sregex_iterator k = words_begin; k != words_end; ++k) {
         std::smatch match = *k;
         std::string match_str = match.str();
-        int pos = int(match_str.find("."));
-        if(pos!=-1) {
+        size_t pos = match_str.find(".");
+        if(pos != -1) {
             std::string new_str = match_str.replace(pos, 1, "_");
             s = s.replace(match.position(), match.str().length(), new_str);
         }
